test(router): split concurrency edge case tests for GH727
