+++ specs/GH727/tech.md
@@ -10,7 +10,7 @@ Link to `product.md`.
 
 ## Current Evidence
 
-At `origin/main@27ac684370e1`, 40 Rust files remain over the U-16 800-line ceiling.
+At `origin/main@d45b427dfb7c`, 39 Rust files remain over the U-16 800-line ceiling.
 The highest-risk files are not all equivalent: some are pure test suites, some are public
 type facades, and some are runtime orchestrators. They need different split patterns.
 
@@ -37,54 +37,54 @@ type facades, and some are runtime orchestrators. They need different split patt
 | P4 | Utility modules | config helpers, net client utils, sync containers | focused utility tests plus concurrency/behavior checks where relevant |
 | P5 | Closure scan | all Rust files | full over-800 scan, final SpecRail update, final PR may close #727 |
 
-## Current Tranche: Bedrock Model Config Projection
+## Current Tranche: Router Concurrency Test Suite
 
 ### Codebase Context
 
 | Area | Files | Current behavior | Why relevant |
 | --- | --- | --- | --- |
-| Bedrock model config facade | `src/core/providers/bedrock/model_config.rs` | Defines Bedrock model family/API/config types, a large legacy `MODEL_CONFIGS` map, lookup helpers, and focused tests. | The file is 1118 lines and duplicates data already represented in `bedrock/catalog`. |
-| Bedrock catalog | `src/core/providers/bedrock/catalog/` | Stores typed catalog entries split by vendor/family and already projects each entry to `ModelConfig`. | This is the existing architectural boundary for Bedrock model metadata. |
-| Bedrock callers | `chat`, `provider`, `transformation`, `model_id`, route spend/token policy | Call `get_model_config` / `get_model_config_for_model_id` or use exported config types. | Public lookup paths and error behavior must remain unchanged. |
+| Router concurrency suite | `src/core/router/tests/concurrency_edge_case_tests.rs` | Contains concurrent selection, model-list swapping, weighted random, EMA, cooldown, and additional edge-case tests. | The file is 1079 lines and already has clear behavior-section boundaries. |
+| Router test module | `src/core/router/tests/mod.rs` | Includes `mod concurrency_edge_case_tests;`. | The root module path should remain stable; the split can happen under a child directory. |
+| Shared helpers | `router_tests::create_test_deployment`, router config/deployment/strategy imports | Shared by multiple test groups. | Keep shared imports in the root test module and have child modules import `super::*`. |
 
 ### Design
 
-1. Keep `src/core/providers/bedrock/model_config.rs` as the public facade for `BedrockModelFamily`, `BedrockApiType`, `ModelConfig`, and lookup helpers.
-2. Replace the hand-written 900+ line `MODEL_CONFIGS` initializer with a projection from `super::catalog::all_entries()`:
-   - key: `entry.model_id`
-   - value: `entry.to_model_config()`
-3. Preserve `get_model_config`, `model_supports_capability`, and `get_all_model_ids` signatures and error behavior.
-4. Update catalog module docs to reflect the new data ownership: the catalog drives the legacy `model_config` facade instead of being only a validation mirror.
-5. Keep existing model_config tests in place; they now verify the public facade behavior.
-6. Keep existing catalog cross-reference tests in place; they continue to verify catalog integrity and projection shape.
-7. Do not edit Bedrock model IDs, pricing values, capabilities, limits, request transformation, provider routing, or model ID parsing.
+1. Keep `src/core/router/tests/concurrency_edge_case_tests.rs` as the root test module.
+2. Move each existing behavior section into a child file under `src/core/router/tests/concurrency_edge_case_tests/`:
+   - `concurrent_selection_tests.rs`
+   - `model_list_swap_tests.rs`
+   - `weighted_random_tests.rs`
+   - `ema_latency_tests.rs`
+   - `cooldown_expiry_tests.rs`
+   - `additional_edge_case_tests.rs`
+3. Root module keeps the doc comment, `#![allow(deprecated)]`, shared imports, and `mod` declarations.
+4. Child modules use `use super::*;` and keep original assertions unchanged.
+5. Do not edit router runtime code or `src/core/router/tests/mod.rs`.
 
 ## Product-to-Test Mapping
 
 | Product invariant | Implementation area | Verification |
 | --- | --- | --- |
-| P1 | `model_config.rs` projection initializer | Focused Bedrock model_config tests pass through the unchanged public helpers. |
-| P2 | `catalog/mod.rs` and existing entries | Catalog tests still pass, including projection and pricing-state invariants. |
-| P3 | Bedrock callers | `cargo check --all-features --locked` compiles existing imports and public re-exports. |
-| P4 | file size | `wc -l src/core/providers/bedrock/model_config.rs src/core/providers/bedrock/catalog/*.rs src/core/providers/bedrock/catalog/entries/*.rs` shows all touched files below 800. |
-| P5 | queue count | `git ls-files '*.rs' | xargs wc -l | awk '$1 > 800 && $2 != "total" { print $1 " " $2 }' | sort -nr` shows the remaining queue no longer includes `model_config.rs`. |
+| P1 | root `concurrency_edge_case_tests.rs` | Keeps shared imports and discovers all child modules. |
+| P2 | child test modules | `cargo test core::router::tests::concurrency_edge_case_tests --lib --all-features` runs the moved tests. |
+| P3 | file size | `wc -l src/core/router/tests/concurrency_edge_case_tests.rs src/core/router/tests/concurrency_edge_case_tests/*.rs` shows all touched files below 800. |
+| P4 | queue count | `git ls-files '*.rs' | xargs wc -l | awk '$1 > 800 && $2 != "total" { print $1 " " $2 }' | sort -nr` shows the remaining queue no longer includes `concurrency_edge_case_tests.rs`. |
 
 ## Risks
 
-- `model_config.rs` and `catalog` reference each other by design: the catalog imports the public config types, while the facade imports `all_entries()` for data projection. Avoid adding runtime calls from catalog entry construction back into `get_model_config`.
-- Catalog tests that previously compared catalog entries to the legacy map become facade-behavior checks after this change; keep model_config unit tests as the direct public surface smoke test.
-- `get_all_model_ids()` order is not specified today because it reads `HashMap` keys; this tranche must not add any order guarantee.
-- This tranche reduces one of the remaining 40 files; the issue remains a tracker after merge.
+- Test names move one module deeper, but the root filter remains stable. PR body should call out the path change.
+- Shared imports must stay available to child modules without introducing a broad test prelude outside this file family.
+- Weighted-random tests are statistical; assertions must not be weakened in this layout-only tranche.
+- This tranche reduces one of the remaining 39 files; the issue remains a tracker after merge.
 
 ## Test Plan
 
 - [ ] `cargo fmt --all -- --check`
-- [ ] `cargo test core::providers::bedrock::model_config --lib --all-features`
-- [ ] `cargo test core::providers::bedrock::catalog --lib --all-features`
+- [ ] `cargo test core::router::tests::concurrency_edge_case_tests --lib --all-features`
 - [ ] `cargo check --all-features --locked`
-- [ ] Line-count proof for `src/core/providers/bedrock/model_config.rs` and touched catalog files
+- [ ] Line-count proof for `src/core/router/tests/concurrency_edge_case_tests.rs` and child files
 
 ## Rollback
 
-Revert the Bedrock model-config projection and `specs/GH727` edits. No migrations
-or runtime config changes are involved.
+Revert the router concurrency test split and `specs/GH727` edits. No migrations
+or runtime code changes are involved.
